+++ blendiff/merge_engine/__init__.py
@@ -1,2 +1,4 @@
 from .merge_engine import MergeEngine
 from .applier import Applier
+
+__all__ = ["MergeEngine", "Applier"]